make points an integer

+++ judge/models/problem.py
@@ -422,7 +422,7 @@ class Problem(models.Model):
         # Site wide voting ban.
         if user.profile.is_banned_from_voting_problem_points:
             return False
-        
+
         # If the user is banned from submitting to the problem.
         if self.banned_users.filter(pk=user.pk).exists():
             return False
@@ -518,7 +518,7 @@ class Solution(models.Model):
 
 
 class ProblemPointsVote(models.Model):
-    points = models.FloatField(
+    points = models.IntegerField(
         verbose_name=_('How much this vote is worth'),
         help_text=_('The amount of points you think this problem deserves.'),
         validators=[
